Refactor component imports and replace deprecated components; update ProblemList data structure
