+++ src/components/AppHeader.vue
@@ -1,5 +1,5 @@
 <script setup lang="ts">
-import HelloWorld from './HelloWorld.vue'
+import HelloWorld from './SubTitle.vue'
 </script>
 
 <template>
